Scripteble object works without animator
