+++ Assets/Global/Animations/Attack/PrefabController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cf4ed520b6c80c74a8329dcf44a6c246
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
