+++ Assets/Global/Animations/Attack/PrefabController/WaffleAnimator.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0cf41981d81c98d49a2778888f5f10bf
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
