+++ Assets/Global/ScriptableObjects/Tails.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b44187384a4a9f42a83d8e8c00838fa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
