+++ Assets/Global/ScriptableObjects/Tails/Waffle.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4573dec063fc903468e001bdc9c403e6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
