--- Assets/Global/Scripts/Player/Attack/Attack.cs
@@ -1,6 +1,0 @@
-using UnityEngine;
-
-public class Attack : MonoBehaviour
-{
-    
-}
