+++ Assets/Global/Scripts/Player/Attacks/Attack.cs
@@ -1,0 +1,3 @@
+using UnityEngine;
+
+public class Attack : MonoBehaviour { }
