+++ Assets/Global/Scripts/Player/Tails.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 694dccb9d6f977f4a833333979136ef6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
