+++ Assets/Global/Scripts/Player/Tails/BaseTail.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6998065adb0cca546b130e890c877371
