+++ Assets/Global/Scripts/Player/Tails/Tail.cs
@@ -6,11 +6,9 @@ public class Tail : MonoBehaviour
 {
     public Player player;
 
-    public List<GameObject> groundCombo = new();
-    public List<GameObject> airCombo = new(); 
-    [HideInInspector]
-    public List<GameObject> currentCombo = new();
-    public Animator animatorTailAttack;
+    public BaseTail currentTail;
+
+    public Animator animator;
 
     public void OnTriggerEnter(Collider Collider)
     {
